refactor: move tips_categories.json to tests/data and fix integration tests
- Move tips_categories.json from project root to tests/data/ folder
- Update test_tips_integration.py to reference new file location
- Fix indentation errors in test methods
- Update README.md examples to reflect new file path
- Improve project organization by separating test data from main code
